added info button to show bucket node card instead of mouse enter event
