+++ src/components/BucketTreeNodeCardButton.svelte
@@ -1,0 +1,223 @@
+<script lang="ts">
+  import { fade } from 'svelte/transition'
+  import Tooltip, { Wrapper } from '@smui/tooltip'
+  import Fa from 'svelte-fa'
+  import { faCircleInfo } from '@fortawesome/free-solid-svg-icons/faCircleInfo'
+  import { clickOutside } from 'svelte-use-click-outside'
+  import type { LayerInfoMetadata, TreeNode } from '$lib/types'
+  import { clean, fetchUrl, getActiveBandIndex, getBase64EncodedUrl, hash } from '$lib/helper'
+  import Popper from '$lib/popper'
+  import { layerMetadata, bucketList } from '$stores'
+  import { TITILER_API_ENDPOINT } from '$lib/constants'
+  import { cloneDeep } from 'lodash-es'
+
+  export let layerInfoMetadata: LayerInfoMetadata = undefined
+  export let node: TreeNode
+
+  const {
+    ref: popperRef,
+    options: popperOptions,
+    content: popperContent,
+  } = new Popper(
+    {
+      placement: 'right',
+      strategy: 'fixed',
+    },
+    [0, 20],
+  ).init()
+
+  const handShowBucketTreeNodeCard = async () => {
+    const layerPathHash = hash(node.path)
+    let metadata: LayerInfoMetadata
+
+    // get existing metadata from store
+    if ($layerMetadata.has(layerPathHash)) {
+      metadata = $layerMetadata.get(layerPathHash)
+    } else {
+      if (node.isStac) {
+        let description = ''
+
+        const bucketStac = $bucketList.find((bucket) => bucket.id === node.path.split('/')[0])
+        const itemsUrl = []
+        itemsUrl.push(bucketStac.url)
+        itemsUrl.push(node.path.split('/')[1])
+        itemsUrl.push('items')
+        itemsUrl.push(node.label)
+        let layerInfo = await fetchUrl(itemsUrl.join('/'))
+
+        if (layerInfo?.properties?.description === undefined) {
+          layerInfo = await fetchUrl(itemsUrl.slice(0, -2).join('/'))
+          description = layerInfo?.description === undefined ? 'N/A' : layerInfo.description
+        } else {
+          description = layerInfo.properties.description
+        }
+
+        const source = layerInfo?.properties?.platform === undefined ? 'N/A' : layerInfo.properties.platform
+        setLayerMetaDataStore(description, source, 'N/A', layerPathHash, true)
+      } else {
+        // get metadata from endpoint
+        const layerURL = new URL(node.url)
+        const infoURI: string = node.isRaster
+          ? `${TITILER_API_ENDPOINT}/info?url=${getBase64EncodedUrl(node.url)}`
+          : `${layerURL.origin}${decodeURIComponent(layerURL.pathname).replace('{z}/{x}/{y}.pbf', 'metadata.json')}${
+              layerURL.search
+            }`
+        const layerInfo = await fetchUrl(infoURI)
+
+        if (node.isRaster) {
+          if (layerInfo?.band_metadata?.length > 0 && !$layerMetadata.has(layerPathHash)) {
+            const bandIndex = getActiveBandIndex(layerInfo)
+            setLayerMetaDataStore(
+              layerInfo.band_metadata[bandIndex][1]['Description'],
+              layerInfo.band_metadata[bandIndex][1]['Source'],
+              layerInfo.band_metadata[bandIndex][1]['Unit'],
+              layerPathHash,
+              true,
+            )
+          }
+        } else {
+          setLayerMetaDataStore(layerInfo.description, layerInfo.source, 'N/A', layerPathHash, true)
+        }
+      }
+
+      metadata = $layerMetadata.get(layerPathHash)
+    }
+
+    if (metadata) {
+      layerInfoMetadata = {
+        description: metadata.description,
+        source: metadata.source,
+        unit: metadata.unit,
+        visible: true,
+      }
+    }
+  }
+
+  const setLayerMetaDataStore = (
+    description: string,
+    source: string,
+    unit: string,
+    layerPathHash: number,
+    visible: boolean,
+  ) => {
+    const metadata = <LayerInfoMetadata>{
+      description,
+      source,
+      unit,
+      visible,
+    }
+    const layerMetadataClone = cloneDeep($layerMetadata)
+    Object.entries(layerMetadataClone).forEach((key) => {
+      const value = layerMetadataClone.get(key)
+      value.visible = false
+      layerMetadataClone.set(key, value)
+    })
+    layerMetadataClone.set(layerPathHash, metadata)
+    $layerMetadata = layerMetadataClone
+    return
+  }
+
+  const handleClose = () => {
+    layerInfoMetadata.visible = false
+  }
+</script>
+
+<div
+  class="icon"
+  alt="Show more detailed information"
+  title="how more detailed information"
+  use:popperRef
+  on:click={() => {
+    handShowBucketTreeNodeCard()
+  }}>
+  <Wrapper>
+    <Fa icon={faCircleInfo} size="sm" />
+    <Tooltip showDelay={0} hideDelay={100} yPos="above">Show more detailed information</Tooltip>
+  </Wrapper>
+</div>
+
+{#if layerInfoMetadata?.visible}
+  <div
+    id="tooltip"
+    data-testid="tooltip"
+    use:popperContent={popperOptions}
+    transition:fade
+    use:clickOutside={handleClose}>
+    <div class="bucket-card">
+      <div class="columns is-vcentered is-mobile">
+        <div class="column is-full">
+          <div class="label">{clean(node.label)}</div>
+          <div class="description">{layerInfoMetadata?.description}</div>
+          <div class="source is-size-6">
+            <span class="has-text-weight-bold">Source: </span>{layerInfoMetadata?.source
+              ? layerInfoMetadata.source
+              : 'N/A'}
+          </div>
+          {#if layerInfoMetadata?.unit}
+            <div class="unit is-size-6">
+              <span class="has-text-weight-bold">Unit: </span>{layerInfoMetadata?.unit ? layerInfoMetadata.unit : 'N/A'}
+            </div>
+          {/if}
+
+          <div class="content is-size-7 tags pt-3">
+            {#if node.tags}
+              {#each Object.values(node.tags) as tag}
+                <span title="tag" style="margin-right: 5px; font-weight: bold;">
+                  <span class="tag is-info is-small is-light">{clean(tag)}</span>
+                </span>
+              {/each}
+            {/if}
+          </div>
+        </div>
+      </div>
+    </div>
+    <div id="arrow" data-popper-arrow />
+  </div>
+{/if}
+
+<style lang="scss">
+  @import '../styles/popper.scss';
+
+  .bucket-card {
+    text-align: justify;
+    text-justify: inter-word;
+    word-wrap: break-word;
+
+    .columns {
+      z-index: 10;
+      position: relative;
+
+      .is-full {
+        .description,
+        .source,
+        .unit {
+          font-weight: normal;
+          color: #000;
+          margin-bottom: 10px;
+
+          @media (prefers-color-scheme: dark) {
+            color: #fff;
+          }
+        }
+
+        .label {
+          border-bottom: 1px solid #ccc;
+          padding-bottom: 5px;
+          margin-bottom: 10px;
+
+          @media (prefers-color-scheme: dark) {
+            color: #fff;
+          }
+        }
+
+        .description {
+          margin-bottom: 15px;
+        }
+      }
+    }
+  }
+
+  #tooltip {
+    max-width: 450px;
+  }
+</style>
